Remove assumptions around IDAM ID type
The type for the IDAM identifier was an implementation detail and should
not have been assumed.

The type is changing soon, to enable testing of the new solution we need
to remove our assumption on the type

+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/controllers/advices/ResourceExceptionHandlerTest.java
@@ -15,7 +15,6 @@ import java.time.LocalDate;
 
 import static java.util.Collections.emptySet;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
@@ -63,7 +62,7 @@ public class ResourceExceptionHandlerTest extends BaseTest {
 
     private ResultActions postClaim(ClaimData claimData) throws Exception {
         return webClient
-            .perform(post("/claims/" + (Long) CLAIMANT_ID)
+            .perform(post("/claims/" + CLAIMANT_ID)
                 .header(HttpHeaders.CONTENT_TYPE, "application/json")
                 .header(HttpHeaders.AUTHORIZATION, "token")
                 .content(jsonMapper.toJson(claimData))
